backend/requirements.txt
Django>=5.0,<6
django-ninja>=1.3.0
django-cors-headers>=4.3.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
